Fix Jena deployment
Deployment of Apache Jena Fuseki was broken because of ReadWriteOnce access mode for PVC. Old pod did not let a new pod to start, so deployment was blocked.

An option to use ReadWriteMany mode is not applicable because Jena is not designed to support multiple instances on top of the same storage. Soo more details in docs: https://jena.apache.org/documentation/tdb/index.html